Изменила views.py использовав forms

+++ todoist/webapp/views.py
@@ -1,7 +1,7 @@
-from django.http import HttpResponseRedirect
 from django.shortcuts import render, redirect, get_object_or_404
 
 from webapp.models import Todo
+from webapp.forms import TodoForm
 
 
 # Create your views here.
@@ -13,15 +13,29 @@ def todo_list_view(request):
 
 def todo_create_view(request):
     if request.method == "GET":
-        return render(request, "create_todo.html")
+        form = TodoForm()
+        return render(request, "create_todo.html", {"form": form})
     else:
-        todo = Todo.objects.create(
-            content=request.POST.get("content"),
-            status=request.POST.get("status"),
-            details=request.POST.get("details"),
-            created_at=request.POST.get("created_at"),
-        )
-        return redirect('todo_view', pk=todo.pk)
+        form = TodoForm(data=request.POST)
+        if form.is_valid():
+            todo = form.save()
+            return redirect('todo_view', pk=todo.pk)
+        else:
+            return render(request, "create_todo.html", {"form": form})
+
+
+def todo_update_view(request, pk):
+    todo = get_object_or_404(Todo, id=pk)
+    if request.method == "GET":
+        form = TodoForm(instance=todo)
+        return render(request, "update_todo.html", {"form": form})
+    else:
+        form = TodoForm(instance=todo, data=request.POST)
+        if form.is_valid():
+            form.save()
+            return redirect('todo_view', pk=todo.pk)
+        else:
+            return render(request, "update_todo.html", {"form": form})
 
 
 def todo_view(request, *args, pk, **kwargs):
@@ -29,10 +43,13 @@ def todo_view(request, *args, pk, **kwargs):
     return render(request, 'todo_view.html', context={'todo': todo})
 
 
-def delete_todo_view(request, *args, pk, **kwargs):
-    todo = Todo.objects.get(id=pk)
-    todo.delete()
-    return redirect('index')
+def delete_todo_view(request, pk):
+    todo = get_object_or_404(Todo, id=pk)
+    if request.method == "GET":
+        return render(request, 'delete_todo.html', {"todo": todo})
+    else:
+        todo.delete()
+        return redirect("index")
